fix jest test path
