+++ __tests__/__fixtures__/multiple-module-no-beforeeach-test-skip-todo-output.js
@@ -12,7 +12,7 @@ module('Acceptance | browse acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-no-beforeeach-test-skip-todo-code.js';
+      '__tests__/__fixtures__/multiple-module-no-beforeeach-test-skip-todo-code.js';
   });
   skip('it renders browse page', async function (assert) {
     await visit(BROWSE_URL);
@@ -27,7 +27,7 @@ module('Acceptance | browse acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-no-beforeeach-test-skip-todo-code.js';
+      '__tests__/__fixtures__/multiple-module-no-beforeeach-test-skip-todo-code.js';
   });
   todo('it renders browse page', async function (assert) {
     await visit(BROWSE_URL);
