+++ __tests__/__fixtures__/multiple-module-one-beforeeach-new-import-code.js
