+++ __tests__/__fixtures__/multiple-module-one-beforeeach-new-import-output.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupApplicationTest } from 'ember-qunit';
-import { click, visit, getTestMetadata } from '@ember/test-helpers';
+import { getTestMetadata } from '@ember/test-helpers';
 const SELECTORS = Object.freeze({
   MOCK_SELECTOR: '[data-test-nav-bar-browse]',
 });
@@ -12,7 +12,9 @@ module('Acceptance | browse acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-no-beforeeach-import-exists-code.js';
+      '__tests__/__fixtures__/multiple-module-one-beforeeach-new-import-code.js';
+    const myConst = 0;
+    noop(); // do some things here
   });
   test('it renders browse page', async function (assert) {
     await visit(BROWSE_URL);
@@ -27,7 +29,9 @@ module('Acceptance | search acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-no-beforeeach-import-exists-code.js';
+      '__tests__/__fixtures__/multiple-module-one-beforeeach-new-import-code.js';
+    const myConst = 0;
+    noop(); // do some things here
   });
   test('it renders search', async function (assert) {
     await visit(BROWSE_URL);
