+++ __tests__/__fixtures__/one-module-no-beforeeach-import-exists-code.js
