+++ __tests__/__fixtures__/one-module-one-beforeeach-import-exists-output.js
@@ -1,6 +1,6 @@
 import { module, test } from 'qunit';
 import { setupApplicationTest } from 'ember-qunit';
-import { getTestMetadata } from '@ember/test-helpers';
+import { click, visit, getTestMetadata } from '@ember/test-helpers';
 const SELECTORS = Object.freeze({
   MOCK_SELECTOR: '[data-test-nav-bar-browse]',
 });
@@ -12,7 +12,7 @@ module('Acceptance | browse acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/one-module-one-beforeeach-new-import-code.js';
+      '__tests__/__fixtures__/one-module-one-beforeeach-import-exists-code.js';
     const myConst = 0;
     noop(); // do some things here
   });
