+++ __tests__/__fixtures__/one-module-one-beforeeach-new-import-output.js
@@ -12,28 +12,11 @@ module('Acceptance | browse acceptance test', function (hooks) {
   hooks.beforeEach(function () {
     let testMetadata = getTestMetadata(this);
     testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-one-beforeeach-new-import-code.js';
+      '__tests__/__fixtures__/one-module-one-beforeeach-new-import-code.js';
     const myConst = 0;
     noop(); // do some things here
   });
   test('it renders browse page', async function (assert) {
-    await visit(BROWSE_URL);
-    assert.dom(SELECTORS.MOCK_SELECTOR).exists();
-  });
-});
-module('Acceptance | search acceptance test', function (hooks) {
-  setupApplicationTest(hooks);
-  some.otherThing(function () {
-    noop();
-  });
-  hooks.beforeEach(function () {
-    let testMetadata = getTestMetadata(this);
-    testMetadata.filePath =
-      '__test__/__fixtures__/multiple-module-one-beforeeach-new-import-code.js';
-    const myConst = 0;
-    noop(); // do some things here
-  });
-  test('it renders search', async function (assert) {
     await visit(BROWSE_URL);
     assert.dom(SELECTORS.MOCK_SELECTOR).exists();
   });
